requirements.txt
# Core dependencies
anthropic>=0.18.0,<1.0.0
langgraph>=0.0.20,<1.0.0
langchain-core>=0.1.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.5.0,<3.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0

# Utilities
jinja2>=3.1.2,<4.0.0

# Type hints and validation
typing-extensions>=4.8.0,<5.0.0

# Optional: For better JSON handling
orjson>=3.9.0,<4.0.0
